Move desktop-client code to the `desktop` package.
