+++ client/src/main/kotlin/io/spine/examples/chatspn/desktop/Main.kt
@@ -23,29 +23,20 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+package io.spine.examples.chatspn.desktop
 
-package io.spine.examples.chatspn
+import androidx.compose.ui.unit.DpSize
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.Window
+import androidx.compose.ui.window.WindowState
+import androidx.compose.ui.window.application
 
-import io.spine.core.UserId
-import io.spine.examples.chatspn.account.UserProfile
-
-/**
- * Provides API to work with users.
- */
-public interface UserProvider {
-
-    /**
-     * Returns user's profile by id.
-     */
-    public fun findUser(id: UserId): UserProfile
-
-    /**
-     * Returns user's profile by email.
-     */
-    public fun findUser(email: String): UserProfile
-
-    /**
-     * Returns the profile of the logged on user.
-     */
-    public fun loggedUser(): UserProfile
+public fun main(): Unit = application {
+    Window(
+        onCloseRequest = ::exitApplication,
+        title = "ChatSPN",
+        state = WindowState(size = DpSize(1000.dp, 600.dp))
+    ) {
+        App()
+    }
 }
